Don't cross compile GetAssemblyVersion
Previously I was cross-compiling the GetAssemblyVersion method to avoid
a dependency on MetadataReader on desktop.  This had the downside of causing
first chance exceptions when run on desktop.
