--- Microsoft.Packaging.Tools/tasks/FileUtilities.netstandard.cs
@@ -1,42 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.IO;
-using System.Reflection.Metadata;
-using System.Reflection.PortableExecutable;
-
-namespace Microsoft.DotNet.Build.Tasks
-{
-    static partial class FileUtilities
-    {
-        private static Version GetAssemblyVersion(string sourcePath)
-        {
-            using (var assemblyStream = new FileStream(sourcePath, FileMode.Open, FileAccess.Read, FileShare.Delete | FileShare.Read))
-            {
-                Version result = null;
-                try
-                {
-                    using (PEReader peReader = new PEReader(assemblyStream, PEStreamOptions.LeaveOpen))
-                    {
-                        if (peReader.HasMetadata)
-                        {
-                            MetadataReader reader = peReader.GetMetadataReader();
-                            if (reader.IsAssembly)
-                            {
-                                result = reader.GetAssemblyDefinition().Version;
-                            }
-                        }
-                    }
-                }
-                catch (BadImageFormatException)
-                {
-                    // not a PE
-                }
-
-                return result;
-            }
-        }
-    }
-}
